Start working on setting up test infrastructure.
