+++ .github/workflows/ci-workflows.yml
@@ -19,6 +19,19 @@ jobs:
         # Code style
         - linux: codestyle
 
+  tests:
+    needs: initial_checks
+    uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@v1
+    with:
+      display: true
+      coverage: codecov
+      libraries: |
+        apt:
+          - '^libxcb.*-dev'
+          - libxkbcommon-x11-0
+          - libegl1-mesa
+          - libhdf5-dev
+
   deploy-examples:
     needs: initial_checks
     if: github.event_name != 'pull_request'
