+++ codecov.yml
@@ -1,0 +1,10 @@
+comment: off
+coverage:
+  status:
+    project:
+      default:
+        target: auto
+        # adjust accordingly based on how flaky your tests are
+        # this allows a 0.001% drop from the previous base commit coverage
+        # basically just to prevent counting zero changes as negative
+        threshold: 0.001%
