+++ glue_ar/common/__init__.py
@@ -1,3 +1,5 @@
 from .marching_cubes import add_isosurface_layer_gltf, add_isosurface_layer_usd  # noqa: F401
 from .scatter import add_scatter_layer_gltf, add_scatter_layer_usd  # noqa: F401
 from .voxels import add_voxel_layers_gltf, add_voxel_layers_usd  # noqa: F401
+from .scatter_export_options import *  # noqa: F401
+from .volume_export_options import *  # noqa: F401
