+++ glue_ar/common/tests/test_scatter_gltf.py
@@ -1,0 +1,5 @@
+from gltflib.gltf import GLTF
+
+
+class TestScatterGLTF:
+    pass
