+++ glue_ar/conftest.py
@@ -1,0 +1,3 @@
+def pytest_configure(config):
+    from glue_ar import setup
+    setup()
