--- setup.cfg
@@ -1,3 +1,0 @@
-[metadata]
-name = glue-ar
-description_file = README.md
