+++ tox.ini
@@ -7,6 +7,8 @@ passenv =
     HOME
 changedir =
     test: .tmp/{envname}
+extras =
+    test: test,qt,jupyter
 commands =
     test: pip freeze
     test: pytest --pyargs glue_ar --cov glue_ar {posargs}
